+++ src/calendar.js
@@ -4,7 +4,7 @@
  * A javascript drag & drop event calendar, that is fully responsive and compatible with all modern browsers.
  * 
  * @file        calendar.js
- * @version     v3.0.0
+ * @version     v2.10.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2023
@@ -13590,7 +13590,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "3.0.0";
+        return "2.10.0";
     };
 
     /**